requirements.txt
python-telegram-bot>=22.0
requests>=2.32.3
telegram>=0.0.1
gunicorn>=23.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.10
python-dotenv
